+++ scripts/plugins.js
@@ -1,0 +1,44 @@
+export async function handleDFredsConvenientEffects(actor, data){
+
+    if(!game?.dfreds?.effectInterface) return;
+
+    const exhaustionLevel = getProperty(data, "data.attributes.exhaustion");
+
+    const exhaustionEffectName = `Exhaustion ${exhaustionLevel}`;
+    const actorUuid = actor.uuid;
+
+    for(let level = 1; level <= 6; level++) {
+        let levelName = `Exhaustion ${level}`;
+        if (levelName !== exhaustionEffectName && game?.dfreds?.effectInterface.hasEffectApplied({
+            effectName: levelName,
+            uuid: actorUuid
+        })){
+            await game?.dfreds?.effectInterface.removeEffect({
+                effectName: levelName,
+                uuid: actorUuid
+            });
+        }
+    }
+
+    await game?.dfreds?.effectInterface.addEffect({ effectName: exhaustionEffectName, uuid: actorUuid });
+
+}
+
+export async function handleCombatUtilityBelt(actor, data){
+
+    if(!game?.cub?.enhancedConditions?.supported) return;
+
+    const exhaustionLevel = getProperty(data, "data.attributes.exhaustion");
+
+    const exhaustionEffectName = `Exhaustion ${exhaustionLevel}`;
+
+    for(let level = 1; level <= 6; level++) {
+        let levelName = `Exhaustion ${level}`;
+        if (levelName !== exhaustionEffectName && game.cub.hasCondition(levelName, actor, { warn: false })){
+            await game.cub.removeCondition(levelName, actor, { warn: false });
+        }
+    }
+
+    await game.cub.addCondition(exhaustionEffectName, actor);
+
+}